requirements.txt
moviepy
openai-whisper
transformers
torch
tkinterdnd2
